Crawler/dags/requirements_crawler.txt
beautifulsoup4
pandas
minio
httpx
psycopg2-binary
